:art: Implement clean go_router navigation system
